a2a_real_llm/order_agent_langraph/pyproject.toml
[project]
name = "order-agent-langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.8",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.8",
    "litellm>=1.72.6.post1",
    "ollama>=0.5.1",
    "openai>=1.88.0",
    "pydantic>=2.11.7",
    "uvicorn>=0.34.3",
]
